gameFunctCalled
This update features a functional game call to a game function inside the Game.dll file. This is done in the new file named callingGameFuncts.hpp. The return value of the call is a C with a squiggle underneath. This update is primarily to store progress. This native function to the game does nothing , but the logic to call game natives/functions is there now and is a huge step in the development process